requirements.txt
aiogram>=3.0.0b7
httpx>=0.23.0
python-dotenv>=1.0.0
